+++ tests/plugins/test_base.py
@@ -29,22 +29,27 @@ from truegaze.plugins.base import BasePlugin
 # Tests for BasePlugin
 class TestBasePlugin(object):
     def test_plugin_properties_name_desc(self):
-        plugin = BasePlugin({}, True, True)
+        plugin = BasePlugin({}, True, True, True)
         assert plugin.name.startswith('Base')
         assert plugin.desc.startswith('Base')
 
-    def test_plugin_properties_support(self):
-        plugin = BasePlugin({}, True, True)
+    def test_plugin_properties(self):
+        plugin = BasePlugin({}, True, True, True)
         assert plugin.supports_android is False
         assert plugin.supports_ios is False
+        assert plugin.supports_online is False
 
-    def test_check_support_is(self):
-        assert BasePlugin({}, False, False).is_os_supported() is False
-        assert BasePlugin({}, False, True).is_os_supported() is False
-        assert BasePlugin({}, True, False).is_os_supported() is False
-        assert BasePlugin({}, True, True).is_os_supported() is False
+    def test_is_os_supported(self):
+        assert BasePlugin({}, False, False, True).is_os_supported() is False
+        assert BasePlugin({}, False, True, True).is_os_supported() is False
+        assert BasePlugin({}, True, False, True).is_os_supported() is False
+        assert BasePlugin({}, True, True, True).is_os_supported() is False
+
+    def test_is_online_testing_supported(self):
+        assert BasePlugin({}, False, False, False).is_online_testing_supported() is False
+        assert BasePlugin({}, False, False, True).is_online_testing_supported() is False
 
     def test_scan_not_implemented(self):
-        plugin = BasePlugin({}, True, True)
+        plugin = BasePlugin({}, True, True, True)
         with pytest.raises(NotImplementedError):
             assert plugin.scan()